pyproject.toml
[project]
name = "windows-use"
version = "0.7.4"
description = "An AI Agent that interacts with Windows OS at GUI level."
authors = [
    { name = "Jeomon George", email = "jeogeoalukka@gmail.com" }
]
readme = "README.md"
license = 'MIT'
license-files = ["LICENSE"]
urls = { homepage = "https://github.com/CursorTouch" }
keywords = ["windows", "agent", "ai", "desktop","ai agent","automation"]
requires-python = ">=3.13"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: User Interfaces",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "fuzzywuzzy>=0.18.0",
    "ipykernel>=6.29.5",
    "pillow>=11.2.1",
    "psutil>=7.0.0",
    "pyautogui>=0.9.54",
    "pydantic>=2.11.7",
    "python-levenshtein>=0.27.1",
    "rich>=14.0.0",
    "tabulate>=0.9.0",
    "termcolor>=3.1.0",
    "twine>=6.1.0",
    "python-dotenv>=1.0.0",
    "groq>=0.29.0",
    "google-genai>=1.45.0",
    "openai>=1.93.0",
    "ollama>=0.5.1",
    "cerebras-cloud-sdk>=1.50.1",
    "anthropic>=0.68.1",
    "mistralai>=1.9.11",
    "pywin32>=311",
    "comtypes>=1.4.14",
    "qrcode>=8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["windows_use"]
